==> scripts/analysis/00_Main.R <==
################################################################################
#
# Main script for analysis of pico contacts
# 
################################################################################

source("./scripts/analysis/01_Load_packages.R")
source("./scripts/analysis/02_Load_data.R")
source("./scripts/analysis/03_Prepare_data.R")
source("./scripts/analysis/04_Analyze_participants.R")
source("./scripts/analysis/05_Analyze_contacts.R")
source("./scripts/analysis/06_Examine_contacts_results.R")
source("./scripts/analysis/07_Plot_main_figure.R")
source("./scripts/analysis/08_Analyze_NGM.R")
source("./scripts/analysis/09_Analyze_workfromhome.R")
source("./scripts/analysis/10_Analyze_contact_type.R")
